+++ lib/world.ts
@@ -6,10 +6,11 @@ import { HasDynamicProperties } from "./dynamic-props.js";
 import { Wrapper } from "./wrapper.js";
 import { MessageType } from "@protobuf-ts/runtime";
 import { IPreferencesContainer } from "./preferences.js";
+import { Vector3, WorldSoundOptions } from "@minecraft/server";
 import * as Prefs from "./preferences.js";
 import * as MC from "@minecraft/server";
 
-export { EntityQueryOptions } from "@minecraft/server";
+export { EntityQueryOptions, WorldSoundOptions } from "@minecraft/server";
 
 export class World extends HasDynamicProperties(Wrapper<MC.World>) implements IPreferencesContainer {
     #isReady: boolean;
@@ -59,6 +60,10 @@ export class World extends HasDynamicProperties(Wrapper<MC.World>) implements IP
         this.setDynamicProperty(
             Prefs.dynamicPropertyId("world"),
             Prefs.encode(ty, prefs));
+    }
+
+    public playSound(soundId: string, location: Vector3, soundOptions?: WorldSoundOptions): void {
+        this.raw.playSound(soundId, location, soundOptions);
     }
 
     #glueEvents(): void {